Reorganize functions
